Fixed page number bug
Fixed page numbers issue. Turned out to be a combination of Docx4J re-serializing certain elements when it didn't need to, and in doing so packaging several instrText and fldChar objects inside a single run. Implemented a solution that undoes these changes when the XML is traversed during PDF generation.